+++ src/components/profile/MyProfile.js
@@ -1,14 +1,35 @@
-import React from 'react';
-import MissionsProfile from './missions/MissionsProfile';
+import { useSelector } from 'react-redux';
+import ProfileRockets from '../ProfileLogic/ProfileRockets';
+import MissionsProfile from '../profile/missions/MissionsProfile';
 
-const MyProfile = () => (
-  <div>
-    <div>
-      <p className="profile-title">
-        My Missions
-      </p>
-      <MissionsProfile />
+const MyProfile = () => {
+  const { rockets, isLoading, error } = useSelector((state) => state.rocket);
+
+  if (isLoading) {
+    return <div>Content is loading...</div>;
+  }
+
+  if (error) {
+    return <div>Something went wrong!!!</div>;
+  }
+
+  return (
+    <div className="container">
+        <div>
+          <p className="profile-title">
+            My Missions
+          </p>
+          <MissionsProfile />
+        </div>
+        <div>
+          <h1>My Rockets</h1>
+          <ul>
+            <ProfileRockets rocketProps={rockets} />
+          </ul>
+        </div>
     </div>
-  </div>
-);
+  );
+};
+
 export default MyProfile;
+
